Refactored to add controller package
