+++ src/main/java/ui/Main.java
@@ -5,8 +5,7 @@ import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.stage.Stage;
-
-import java.io.IOException;
+import ui.controller.BaseController;
 
 /**
  * Created by Ari on 2/4/17.
